+++ share/gui.h
@@ -98,7 +98,7 @@ void gui_paint(int);
 void gui_pulse(int, float);
 void gui_timer(int, float);
 int  gui_point(int, int, int);
-int  gui_stick(int, int, int);
+int  gui_stick(int, int, float);
 void gui_stuck(void);
 int  gui_click(void);
 void gui_focus(int);
